feat: consolidate admin UX into the webapp, mobile-first redesign, i18n overhaul
Removes the cluttered /admin_* Telegram command menu (only /admin and /sync
remain registered) now that internal/adminops backs both the bot and the
webapp identically; adds a "send test broadcast to me" endpoint to close the
one capability gap left by that removal.

Every admin mutating action (topup, reset devices, reset traffic, enable/
disable, extend, trial toggle) now notifies the affected customer through
the translation system in their own language, closing gaps where 4 of 7
actions were silent and 3 were hardcoded Russian. Fixes accumulated
translation inconsistencies (EN guillemets, %d/%.1f traffic formatting,
bold-header style, an informal-register slip in two Tribute strings).

web/admin gets a mobile-first pass: off-canvas sidebar drawer below 720px,
card-list fallback for data tables, 44px touch targets, safe-area insets,
and native Telegram BackButton support.

Adds two new admin capabilities: a query-time activity feed (signups,
purchases, referral bonuses, admin actions merged by timestamp, no new
write path) and one-way admin-to-customer messaging (migration 000012 adds
admin_audit_log.param_text to keep sent text in the audit trail).

+++ internal/notification/traffic.go
@@ -72,8 +72,8 @@ func (s *TrafficWarningService) CheckAndNotify() error {
 			continue
 		}
 
-		remainingGB := (u.TrafficLimitBytes - u.UserTraffic.UsedTrafficBytes) / config.BytesInGigabyte()
-		totalGB := u.TrafficLimitBytes / config.BytesInGigabyte()
+		remainingGB := float64(u.TrafficLimitBytes-u.UserTraffic.UsedTrafficBytes) / float64(config.BytesInGigabyte())
+		totalGB := float64(u.TrafficLimitBytes) / float64(config.BytesInGigabyte())
 		text := fmt.Sprintf(s.tm.GetText(customer.Language, "traffic_warning"), remainingGB, totalGB)
 
 		var rows [][]models.InlineKeyboardButton
